rewrite BasicDataLoader based on dev_nc_blob (#496)
* change dataloader

* make basic_data_loader_kernel.cpp readable

* debug

* debug dataloader

* change something due to review

* do some changes

* refactor to reduce code

* remove comment

* refactor the dataloader kernel

* debug

* debug a runtime bug

* remove fillblobwithzero method

* changes based on reviews

* debug and make it runnable

* change the declarations due to zyeric's change request

* changes due to willzhang's change request

* make it able to compile

* change the type of blob header and move the position of DataLoadStatus

* change the declaration of function ReadColFromBufferToOutBlob()

* change the wrong use of memcpy() and memset()

* add 0 after the end of each line

* refactor ReadOneDataContent() function

* revert a wrong change

* remove unnecessary CHECK()

* rename and move some declaration

* bug for compiles

+++ oneflow/core/register/blob_desc.h
@@ -9,8 +9,8 @@
 namespace oneflow {
 
 struct BlobHeader {
-  int64_t col_id;
-  int64_t max_col_id;
+  int32_t col_id;
+  int32_t max_col_id;
 };
 
 class BlobDesc final {
